// src/app/interfaces/persons.ts
export interface Persons {
  id?: number;
  name: string;
  email?: string;
  phone?: string;
  role?: string;
  customer_tenant_id?: string;
  participant_id?: string;
  customer_id?: string;
  customer_name?: string;
  project_id?: string;
  project_name?: string;
  value?: number;
  label?: string;
  created_at?: any;
  updated_at?: any;
}